requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-google-genai
pypdf
python-dotenv